# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Card {
  id: string;
  name: string;
  country: string;
  position: string;
  rarity: string;
  weight: number;
  image: string;
  flag: string;
}

export interface PullResult {
  card: Card;
}

export interface Pull10Result {
  cards: Card[];
}

export interface LeaderboardEntry {
  username: string;
  total: number;
  legendaries: number;
}

export interface FeedEntry {
  username: string;
  card: Card;
  timestamp: number;
}

export interface UserStats {
  pulls: number;
  owned: number;
  legendaries: number;
  epics: number;
  rares: number;
  commons: number;
}

export type PullCardParams = {
username?: string;
};

export type PullTenCardsParams = {
username?: string;
};