# lib/api-zod/src/generated/types/card.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Card {
  id: string;
  name: string;
  country: string;
  position: string;
  rarity: string;
  weight: number;
  image: string;
  flag: string;
}