# lib/api-zod/src/generated/types/feedEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Card } from './card';

export interface FeedEntry {
  username: string;
  card: Card;
  timestamp: number;
}