# lib/api-zod/src/generated/types/pull10Result.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Card } from './card';

export interface Pull10Result {
  cards: Card[];
}