# lib/api-zod/src/generated/types/userStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserStats {
  pulls: number;
  owned: number;
  legendaries: number;
  epics: number;
  rares: number;
  commons: number;
}